+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # 2.0.0-alpha.4
+* Remove `package:json_god`.
+* Make `WebSocketContext` take any `StreamChannel`.
 * Strong typing updates.
 
 # 2.0.0-alpha.3